Gone back to using Double for the order quantity when getting a ship estimate, it's pulled from the database in a lot of places so it can't always be a BigDecimal for now.
git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@590578 13f79535-47bb-0310-9956-ffa450edef68

+++ applications/order/src/org/ofbiz/order/order/OrderReadHelper.java
@@ -1222,7 +1222,7 @@ public class OrderReadHelper {
     public Map getItemInfoMap(GenericValue item) {
         Map itemInfo = FastMap.newInstance();
         itemInfo.put("productId", item.getString("productId"));
-        itemInfo.put("quantity", getOrderItemQuantity(item));
+        itemInfo.put("quantity", Double.valueOf(getOrderItemQuantity(item).doubleValue()));
         itemInfo.put("weight", new Double(this.getItemWeight(item).doubleValue()));
         itemInfo.put("size",  new Double(this.getItemSize(item)));
         itemInfo.put("piecesIncluded", new Long(this.getItemPiecesIncluded(item)));
